requirements.txt
pywebview
transformers
torch
